update: move useTitle and extend useScroll with scroll wrapper
