+++ src/App.js
@@ -6,10 +6,9 @@ import { useTranslation } from "react-i18next";
 import {
   AuthGuard,
   SplashScreen,
-  ScrollWrapper,
   VerificationGuard,
 } from "components";
-import { useLanguage } from "hooks";
+import { useLanguage, useScroll } from "hooks";
 
 const Sync = lazy(() => import("pages/Sync"));
 const Contact = lazy(() => import("pages/Contact"));
@@ -53,7 +52,7 @@ const BetaTesting = lazy(() => import("pages/BetaTesting"));
 const App = () => {
   const { t } = useTranslation();
   const { LanguageWrapper } = useLanguage();
-
+  const { ScrollWrapper } = useScroll();
   return (
     <Fragment>
       <Toaster
