+++ src/components/index.js
@@ -5,11 +5,9 @@ export * from "./MainNavbar";
 export * from "./Footer";
 export * from "./Loader";
 export * from "./PageAnimationWrapper";
-export * from "./useTitle";
 export * from "./guards";
 export * from "./NavLinks";
 export * from "./Alert";
 export * from "./QRCode";
-export * from "./ScrollWrapper";
 export * from "./Announcement";
 export * from "./form";
