+++ src/hooks/index.js
@@ -1,3 +1,4 @@
+export * from "./useTitle";
 export * from "./useScroll";
 export * from "./useCountDown";
 export * from "./useLanguage";
