+++ src/hooks/useScroll.js
@@ -1,4 +1,6 @@
 import { useEffect, useState, useCallback } from "react";
+import { useLocation } from "react-router-dom";
+import PropTypes from "prop-types";
 
 export const useScroll = () => {
   const [scrolled, setScrolled] = useState(false);
@@ -18,5 +20,34 @@ export const useScroll = () => {
     };
   }, [handleScroll]);
 
-  return scrolled;
+
+  //ref : https://www.kindacode.com/article/react-router-dom-scroll-to-top-on-route-change/
+  //ref : https://github.com/remix-run/react-router/issues/394
+
+
+  const ScrollWrapper = ({ children }) => {
+    const location = useLocation();
+    useEffect(() => {
+      const { hash } = location;
+      if (hash) {
+        const element = document.querySelector(hash);
+        if (element) {
+          element.scrollIntoView();
+        }
+      } else {
+        window.scrollTo(0, 0);
+      }
+    }, [location]);
+
+    return children;
+  };
+
+  ScrollWrapper.propTypes = {
+    children: PropTypes.node,
+  };
+
+  return {
+    ScrollWrapper,
+    scrolled
+  };
 };
