+++ src/pages/password-reset/PasswordReset.js
@@ -13,8 +13,9 @@ import {
   Button,
   PasswordInput,
   PageAnimationWrapper,
-  useTitle,
 } from "components";
+import { useTitle } from "hooks";
+
 
 const PasswordReset = () => {
   const { t } = useTranslation();
