+++ src/pages/settings/AccountDeletion.js
@@ -14,10 +14,10 @@ import {
   Label,
   Button,
   Loader,
-  useTitle,
   FormGroup,
   PasswordInput,
 } from "components";
+import { useTitle } from "hooks";
 
 const AccountDeletion = () => {
   const { t } = useTranslation();
